backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.19
mysql-connector-python==9.2.0
pydantic==2.11.3
pydantic-settings==2.0.3
jinja2==3.1.6
psutil==5.9.5
neo4j==5.14.1
py2neo==2021.2.4
numpy==2.2.4
openai==1.73.0
tenacity==9.1.2
dotenv==0.9.9
python-dotenv==1.1.0
tiktoken==0.9.0
networkx==3.4.2
pydantic==2.11.3
httpx==0.28.1